feature: Remove events dependency on gateway
