+++ lib/src/gateway/shard.dart
@@ -7,7 +7,7 @@ import 'package:nyxx/src/builders/voice.dart';
 import 'package:nyxx/src/client.dart';
 import 'package:nyxx/src/gateway/message.dart';
 import 'package:nyxx/src/gateway/shard_runner.dart';
-import 'package:nyxx/src/models/gateway/event.dart';
+import 'package:nyxx/src/models/events/event.dart';
 import 'package:nyxx/src/models/gateway/opcode.dart';
 import 'package:nyxx/src/models/snowflake.dart';
