+++ lib/src/models/events/application_command.dart
@@ -1,5 +1,5 @@
 import 'package:nyxx/src/models/commands/application_command_permissions.dart';
-import 'package:nyxx/src/models/gateway/event.dart';
+import 'package:nyxx/src/models/events/event.dart';
 
 /// {@template application_command_permissions_update_event}
 /// Emitted when the permissions for an application command are updated.
@@ -13,5 +13,5 @@ class ApplicationCommandPermissionsUpdateEvent extends DispatchEvent {
 
   /// {@macro application_command_permissions_update_event}
   /// @nodoc
-  ApplicationCommandPermissionsUpdateEvent({required super.gateway, required this.permissions, required this.oldPermissions});
+  ApplicationCommandPermissionsUpdateEvent({required super.client, required this.permissions, required this.oldPermissions});
 }
