+++ lib/src/models/events/auto_moderation.dart
@@ -1,6 +1,6 @@
 import 'package:nyxx/src/models/channel/channel.dart';
 import 'package:nyxx/src/models/channel/text_channel.dart';
-import 'package:nyxx/src/models/gateway/event.dart';
+import 'package:nyxx/src/models/events/event.dart';
 import 'package:nyxx/src/models/guild/auto_moderation.dart';
 import 'package:nyxx/src/models/guild/guild.dart';
 import 'package:nyxx/src/models/guild/member.dart';
@@ -17,7 +17,7 @@ class AutoModerationRuleCreateEvent extends DispatchEvent {
 
   /// {@macro auto_moderation_rule_create_event}
   /// @nodoc
-  AutoModerationRuleCreateEvent({required super.gateway, required this.rule});
+  AutoModerationRuleCreateEvent({required super.client, required this.rule});
 }
 
 /// {@template auto_moderation_rule_update_event}
@@ -32,7 +32,7 @@ class AutoModerationRuleUpdateEvent extends DispatchEvent {
 
   /// {@macro auto_moderation_rule_update_event}
   /// @nodoc
-  AutoModerationRuleUpdateEvent({required super.gateway, required this.oldRule, required this.rule});
+  AutoModerationRuleUpdateEvent({required super.client, required this.oldRule, required this.rule});
 }
 
 /// {@template auto_moderation_rule_delete_event}
@@ -44,7 +44,7 @@ class AutoModerationRuleDeleteEvent extends DispatchEvent {
 
   /// {@macro auto_moderation_rule_delete_event}
   /// @nodoc
-  AutoModerationRuleDeleteEvent({required super.gateway, required this.rule});
+  AutoModerationRuleDeleteEvent({required super.client, required this.rule});
 }
 
 /// {@template auto_moderation_action_execution_event}
@@ -87,7 +87,7 @@ class AutoModerationActionExecutionEvent extends DispatchEvent {
   /// {@macro auto_moderation_action_execution_event}
   /// @nodoc
   AutoModerationActionExecutionEvent({
-    required super.gateway,
+    required super.client,
     required this.guildId,
     required this.action,
     required this.ruleId,
@@ -102,19 +102,19 @@ class AutoModerationActionExecutionEvent extends DispatchEvent {
   });
 
   /// The guild the rule was triggered in.
-  PartialGuild get guild => gateway.client.guilds[guildId];
+  PartialGuild get guild => client.guilds[guildId];
 
   /// The rule that was triggered.
   PartialAutoModerationRule get rule => guild.autoModerationRules[ruleId];
 
   /// The user that triggered the rule.
-  PartialUser get user => gateway.client.users[userId];
+  PartialUser get user => client.users[userId];
 
   /// The member that triggered the rule.
   PartialMember get member => guild.members[userId];
 
   /// The channel in which the rule was triggered.
-  PartialChannel? get channel => channelId == null ? null : gateway.client.channels[channelId!];
+  PartialChannel? get channel => channelId == null ? null : client.channels[channelId!];
 
   /// The message that triggered the rule.
   PartialMessage? get message => messageId == null ? null : (channel as PartialTextChannel?)?.messages[messageId!];
