+++ lib/src/models/events/entitlement.dart
@@ -1,5 +1,5 @@
 import 'package:nyxx/src/models/entitlement.dart';
-import 'package:nyxx/src/models/gateway/event.dart';
+import 'package:nyxx/src/models/events/event.dart';
 
 /// {@template entitlement_create_event}
 /// Emitted when an entitlement is created.
@@ -10,7 +10,7 @@ class EntitlementCreateEvent extends DispatchEvent {
 
   /// {@macro entitlement_create_event}
   /// @nodoc
-  EntitlementCreateEvent({required super.gateway, required this.entitlement});
+  EntitlementCreateEvent({required super.client, required this.entitlement});
 }
 
 /// {@template entitlement_update_event}
@@ -25,7 +25,7 @@ class EntitlementUpdateEvent extends DispatchEvent {
 
   /// {@macro entitlement_update_event}
   /// @nodoc
-  EntitlementUpdateEvent({required super.gateway, required this.entitlement, required this.oldEntitlement});
+  EntitlementUpdateEvent({required super.client, required this.entitlement, required this.oldEntitlement});
 }
 
 /// {@template entitlement_delete_event}
@@ -40,5 +40,5 @@ class EntitlementDeleteEvent extends DispatchEvent {
 
   /// {@macro entitlement_delete_event}
   /// @nodoc
-  EntitlementDeleteEvent({required super.gateway, required this.entitlement, required this.deletedEntitlement});
+  EntitlementDeleteEvent({required super.client, required this.entitlement, required this.deletedEntitlement});
 }
