+++ lib/src/models/events/guild.dart
@@ -2,8 +2,8 @@ import 'package:nyxx/src/models/channel/guild_channel.dart';
 import 'package:nyxx/src/models/channel/stage_instance.dart';
 import 'package:nyxx/src/models/channel/thread.dart';
 import 'package:nyxx/src/models/emoji.dart';
-import 'package:nyxx/src/models/gateway/event.dart';
-import 'package:nyxx/src/models/gateway/events/presence.dart';
+import 'package:nyxx/src/models/events/event.dart';
+import 'package:nyxx/src/models/events/presence.dart';
 import 'package:nyxx/src/models/guild/audit_log.dart';
 import 'package:nyxx/src/models/guild/guild.dart';
 import 'package:nyxx/src/models/guild/member.dart';
@@ -23,7 +23,7 @@ class UnavailableGuildCreateEvent extends DispatchEvent {
 
   /// {@macro unavailable_guild_create_event}
   /// @nodoc
-  UnavailableGuildCreateEvent({required super.gateway, required this.guild});
+  UnavailableGuildCreateEvent({required super.client, required this.guild});
 }
 
 /// {@template guild_create_event}
@@ -66,7 +66,7 @@ class GuildCreateEvent extends DispatchEvent implements UnavailableGuildCreateEv
   /// {@macro guild_create_event}
   /// @nodoc
   GuildCreateEvent({
-    required super.gateway,
+    required super.client,
     required this.guild,
     required this.joinedAt,
     required this.isLarge,
@@ -93,7 +93,7 @@ class GuildUpdateEvent extends DispatchEvent {
 
   /// {@macro guild_update_event}
   /// @nodoc
-  GuildUpdateEvent({required super.gateway, required this.oldGuild, required this.guild});
+  GuildUpdateEvent({required super.client, required this.oldGuild, required this.guild});
 }
 
 /// {@template guild_delete_event}
@@ -111,7 +111,7 @@ class GuildDeleteEvent extends DispatchEvent {
 
   /// {@macro guild_delete_event}
   /// @nodoc
-  GuildDeleteEvent({required super.gateway, required this.guild, required this.isUnavailable, required this.deletedGuild});
+  GuildDeleteEvent({required super.client, required this.guild, required this.isUnavailable, required this.deletedGuild});
 }
 
 /// {@template guild_audit_log_create_event}
@@ -126,10 +126,10 @@ class GuildAuditLogCreateEvent extends DispatchEvent {
 
   /// {@macro guild_audit_log_create_event}
   /// @nodoc
-  GuildAuditLogCreateEvent({required super.gateway, required this.entry, required this.guildId});
+  GuildAuditLogCreateEvent({required super.client, required this.entry, required this.guildId});
 
   /// The guild in which the entry was created.
-  PartialGuild get guild => gateway.client.guilds[guildId];
+  PartialGuild get guild => client.guilds[guildId];
 }
 
 /// {@template guild_ban_add_event}
@@ -144,10 +144,10 @@ class GuildBanAddEvent extends DispatchEvent {
 
   /// {@macro guild_ban_add_event}
   /// @nodoc
-  GuildBanAddEvent({required super.gateway, required this.guildId, required this.user});
+  GuildBanAddEvent({required super.client, required this.guildId, required this.user});
 
   /// The guild in which the user was banned.
-  PartialGuild get guild => gateway.client.guilds[guildId];
+  PartialGuild get guild => client.guilds[guildId];
 }
 
 /// {@template guild_ban_remove_event}
@@ -162,10 +162,10 @@ class GuildBanRemoveEvent extends DispatchEvent {
 
   /// {@macro guild_ban_remove_event}
   /// @nodoc
-  GuildBanRemoveEvent({required super.gateway, required this.guildId, required this.user});
+  GuildBanRemoveEvent({required super.client, required this.guildId, required this.user});
 
   /// The guild in which the user was unbanned.
-  PartialGuild get guild => gateway.client.guilds[guildId];
+  PartialGuild get guild => client.guilds[guildId];
 }
 
 /// {@template guild_emojis_update_event}
@@ -180,10 +180,10 @@ class GuildEmojisUpdateEvent extends DispatchEvent {
 
   /// {@macro guild_emojis_update_event}
   /// @nodoc
-  GuildEmojisUpdateEvent({required super.gateway, required this.guildId, required this.emojis});
+  GuildEmojisUpdateEvent({required super.client, required this.guildId, required this.emojis});
 
   /// The guild in which emojis were updated.
-  PartialGuild get guild => gateway.client.guilds[guildId];
+  PartialGuild get guild => client.guilds[guildId];
 }
 
 /// {@template guild_stickers_update_event}
@@ -198,10 +198,10 @@ class GuildStickersUpdateEvent extends DispatchEvent {
 
   /// {@macro guild_stickers_update_event}
   /// @nodoc
-  GuildStickersUpdateEvent({required super.gateway, required this.guildId, required this.stickers});
+  GuildStickersUpdateEvent({required super.client, required this.guildId, required this.stickers});
 
   /// The guild in which the stickers were updated.
-  PartialGuild get guild => gateway.client.guilds[guildId];
+  PartialGuild get guild => client.guilds[guildId];
 }
 
 /// {@template guild_integrations_update_event}
@@ -213,10 +213,10 @@ class GuildIntegrationsUpdateEvent extends DispatchEvent {
 
   /// {@macro guild_integrations_update_event}
   /// @nodoc
-  GuildIntegrationsUpdateEvent({required super.gateway, required this.guildId});
+  GuildIntegrationsUpdateEvent({required super.client, required this.guildId});
 
   /// The guild in which the integrations were updated.
-  PartialGuild get guild => gateway.client.guilds[guildId];
+  PartialGuild get guild => client.guilds[guildId];
 }
 
 /// {@template guild_member_add_event}
@@ -231,10 +231,10 @@ class GuildMemberAddEvent extends DispatchEvent {
 
   /// {@macro guild_member_add_event}
   /// @nodoc
-  GuildMemberAddEvent({required super.gateway, required this.guildId, required this.member});
+  GuildMemberAddEvent({required super.client, required this.guildId, required this.member});
 
   /// The guild in which the member was added.
-  PartialGuild get guild => gateway.client.guilds[guildId];
+  PartialGuild get guild => client.guilds[guildId];
 }
 
 /// {@template guild_member_remove_event}
@@ -252,10 +252,10 @@ class GuildMemberRemoveEvent extends DispatchEvent {
 
   /// {@macro guild_member_remove_event}
   /// @nodoc
-  GuildMemberRemoveEvent({required super.gateway, required this.guildId, required this.user, required this.removedMember});
+  GuildMemberRemoveEvent({required super.client, required this.guildId, required this.user, required this.removedMember});
 
   /// The guild in which the member was removed.
-  PartialGuild get guild => gateway.client.guilds[guildId];
+  PartialGuild get guild => client.guilds[guildId];
 }
 
 /// {@template guild_member_update_event}
@@ -273,10 +273,10 @@ class GuildMemberUpdateEvent extends DispatchEvent {
 
   /// {@macro guild_member_update_event}
   /// @nodoc
-  GuildMemberUpdateEvent({required super.gateway, required this.oldMember, required this.member, required this.guildId});
+  GuildMemberUpdateEvent({required super.client, required this.oldMember, required this.member, required this.guildId});
 
   /// The guild in which the member was updated.
-  PartialGuild get guild => gateway.client.guilds[guildId];
+  PartialGuild get guild => client.guilds[guildId];
 }
 
 /// {@template guild_members_chunk_event}
@@ -307,7 +307,7 @@ class GuildMembersChunkEvent extends DispatchEvent {
   /// {@macro guild_members_chunk_event}
   /// @nodoc
   GuildMembersChunkEvent({
-    required super.gateway,
+    required super.client,
     required this.guildId,
     required this.members,
     required this.chunkIndex,
@@ -318,7 +318,7 @@ class GuildMembersChunkEvent extends DispatchEvent {
   });
 
   /// The guild members are being sent from.
-  PartialGuild get guild => gateway.client.guilds[guildId];
+  PartialGuild get guild => client.guilds[guildId];
 }
 
 /// {@template guild_role_create_event}
@@ -333,10 +333,10 @@ class GuildRoleCreateEvent extends DispatchEvent {
 
   /// {@macro guild_role_create_event}
   /// @nodoc
-  GuildRoleCreateEvent({required super.gateway, required this.guildId, required this.role});
+  GuildRoleCreateEvent({required super.client, required this.guildId, required this.role});
 
   /// The guild in which the role was created.
-  PartialGuild get guild => gateway.client.guilds[guildId];
+  PartialGuild get guild => client.guilds[guildId];
 }
 
 /// {@template guild_role_update_event}
@@ -354,10 +354,10 @@ class GuildRoleUpdateEvent extends DispatchEvent {
 
   /// {@macro guild_role_update_event}
   /// @nodoc
-  GuildRoleUpdateEvent({required super.gateway, required this.guildId, required this.oldRole, required this.role});
+  GuildRoleUpdateEvent({required super.client, required this.guildId, required this.oldRole, required this.role});
 
   /// The guild in which the role was updated.
-  PartialGuild get guild => gateway.client.guilds[guildId];
+  PartialGuild get guild => client.guilds[guildId];
 }
 
 /// {@template guild_role_delete_event}
@@ -375,10 +375,10 @@ class GuildRoleDeleteEvent extends DispatchEvent {
 
   /// {@macro guild_role_delete_event}
   /// @nodoc
-  GuildRoleDeleteEvent({required super.gateway, required this.roleId, required this.guildId, required this.deletedRole});
+  GuildRoleDeleteEvent({required super.client, required this.roleId, required this.guildId, required this.deletedRole});
 
   /// The guild in which the role was deleted.
-  PartialGuild get guild => gateway.client.guilds[guildId];
+  PartialGuild get guild => client.guilds[guildId];
 }
 
 /// {@template guild_scheduled_event_create_event}
@@ -390,7 +390,7 @@ class GuildScheduledEventCreateEvent extends DispatchEvent {
 
   /// {@macro guild_scheduled_event_create_event}
   /// @nodoc
-  GuildScheduledEventCreateEvent({required super.gateway, required this.event});
+  GuildScheduledEventCreateEvent({required super.client, required this.event});
 }
 
 /// {@template guild_scheduled_event_update_event}
@@ -405,7 +405,7 @@ class GuildScheduledEventUpdateEvent extends DispatchEvent {
 
   /// {@macro guild_scheduled_event_update_event}
   /// @nodoc
-  GuildScheduledEventUpdateEvent({required super.gateway, required this.oldEvent, required this.event});
+  GuildScheduledEventUpdateEvent({required super.client, required this.oldEvent, required this.event});
 }
 
 /// {@template guild_scheduled_event_delete_event}
@@ -417,7 +417,7 @@ class GuildScheduledEventDeleteEvent extends DispatchEvent {
 
   /// {@macro guild_scheduled_event_delete_event}
   /// @nodoc
-  GuildScheduledEventDeleteEvent({required super.gateway, required this.event});
+  GuildScheduledEventDeleteEvent({required super.client, required this.event});
 }
 
 /// {@template guild_scheduled_event_user_add_event}
@@ -435,16 +435,16 @@ class GuildScheduledEventUserAddEvent extends DispatchEvent {
 
   /// {@macro guild_scheduled_event_user_add_event}
   /// @nodoc
-  GuildScheduledEventUserAddEvent({required super.gateway, required this.scheduledEventId, required this.userId, required this.guildId});
+  GuildScheduledEventUserAddEvent({required super.client, required this.scheduledEventId, required this.userId, required this.guildId});
 
   /// The guild that the scheduled event is in.
-  PartialGuild get guild => gateway.client.guilds[guildId];
+  PartialGuild get guild => client.guilds[guildId];
 
   /// The scheduled event.
   PartialScheduledEvent get scheduledEvent => guild.scheduledEvents[scheduledEventId];
 
   /// The user that was added.
-  PartialUser get user => gateway.client.users[userId];
+  PartialUser get user => client.users[userId];
 
   /// The member that was added.
   PartialMember get member => guild.members[userId];
@@ -465,16 +465,16 @@ class GuildScheduledEventUserRemoveEvent extends DispatchEvent {
 
   /// {@macro guild_scheduled_event_user_remove_event}
   /// @nodoc
-  GuildScheduledEventUserRemoveEvent({required super.gateway, required this.scheduledEventId, required this.userId, required this.guildId});
+  GuildScheduledEventUserRemoveEvent({required super.client, required this.scheduledEventId, required this.userId, required this.guildId});
 
   /// The guild that the scheduled event is in.
-  PartialGuild get guild => gateway.client.guilds[guildId];
+  PartialGuild get guild => client.guilds[guildId];
 
   /// The scheduled event.
   PartialScheduledEvent get scheduledEvent => guild.scheduledEvents[scheduledEventId];
 
   /// The user that was removed.
-  PartialUser get user => gateway.client.users[userId];
+  PartialUser get user => client.users[userId];
 
   /// The member that was removed.
   PartialMember get member => guild.members[userId];
