+++ lib/src/models/events/integration.dart
@@ -1,5 +1,5 @@
 import 'package:nyxx/src/models/application.dart';
-import 'package:nyxx/src/models/gateway/event.dart';
+import 'package:nyxx/src/models/events/event.dart';
 import 'package:nyxx/src/models/guild/guild.dart';
 import 'package:nyxx/src/models/guild/integration.dart';
 import 'package:nyxx/src/models/snowflake.dart';
@@ -16,10 +16,10 @@ class IntegrationCreateEvent extends DispatchEvent {
 
   /// {@macro integration_create_event}
   /// @nodoc
-  IntegrationCreateEvent({required super.gateway, required this.guildId, required this.integration});
+  IntegrationCreateEvent({required super.client, required this.guildId, required this.integration});
 
   /// The guild the integration was created in.
-  PartialGuild get guild => gateway.client.guilds[guildId];
+  PartialGuild get guild => client.guilds[guildId];
 }
 
 /// {@template integration_update_event}
@@ -37,10 +37,10 @@ class IntegrationUpdateEvent extends DispatchEvent {
 
   /// {@macro integration_update_event}
   /// @nodoc
-  IntegrationUpdateEvent({required super.gateway, required this.guildId, required this.oldIntegration, required this.integration});
+  IntegrationUpdateEvent({required super.client, required this.guildId, required this.oldIntegration, required this.integration});
 
   /// The guild the integration was updated in.
-  PartialGuild get guild => gateway.client.guilds[guildId];
+  PartialGuild get guild => client.guilds[guildId];
 }
 
 /// {@template integration_delete_event}
@@ -61,11 +61,11 @@ class IntegrationDeleteEvent extends DispatchEvent {
 
   /// {@macro integration_delete_event}
   /// @nodoc
-  IntegrationDeleteEvent({required super.gateway, required this.id, required this.guildId, required this.applicationId, required this.deletedIntegration});
+  IntegrationDeleteEvent({required super.client, required this.id, required this.guildId, required this.applicationId, required this.deletedIntegration});
 
   /// The guild the integration was deleted from.
-  PartialGuild get guild => gateway.client.guilds[guildId];
+  PartialGuild get guild => client.guilds[guildId];
 
   /// The application associated with the integration.
-  PartialApplication? get application => applicationId == null ? null : gateway.client.applications[applicationId!];
+  PartialApplication? get application => applicationId == null ? null : client.applications[applicationId!];
 }
