+++ lib/src/models/events/soundboard.dart
@@ -1,4 +1,4 @@
-import 'package:nyxx/src/models/gateway/event.dart';
+import 'package:nyxx/src/models/events/event.dart';
 import 'package:nyxx/src/models/guild/guild.dart';
 import 'package:nyxx/src/models/snowflake.dart';
 import 'package:nyxx/src/models/soundboard/soundboard.dart';
@@ -12,7 +12,7 @@ class SoundboardSoundCreateEvent extends DispatchEvent {
 
   /// {@macro soundboard_sound_create_event}
   /// @nodoc
-  SoundboardSoundCreateEvent({required super.gateway, required this.sound});
+  SoundboardSoundCreateEvent({required super.client, required this.sound});
 }
 
 /// {@template soundboard_sound_update_event}
@@ -27,7 +27,7 @@ class SoundboardSoundUpdateEvent extends DispatchEvent {
 
   /// {@macro soundboard_sound_update_event}
   /// @nodoc
-  SoundboardSoundUpdateEvent({required super.gateway, required this.sound, required this.oldSound});
+  SoundboardSoundUpdateEvent({required super.client, required this.sound, required this.oldSound});
 }
 
 /// {@template soundboard_sound_delete_event}
@@ -45,9 +45,9 @@ class SoundboardSoundDeleteEvent extends DispatchEvent {
 
   /// {@macro soundboard_sound_delete_event}
   /// @nodoc
-  SoundboardSoundDeleteEvent({required super.gateway, required this.sound, required this.guildId, required this.soundId});
+  SoundboardSoundDeleteEvent({required super.client, required this.sound, required this.guildId, required this.soundId});
 
-  PartialGuild get guild => gateway.client.guilds[guildId];
+  PartialGuild get guild => client.guilds[guildId];
 }
 
 /// {@template soundboard_sounds_update_event}
@@ -65,5 +65,5 @@ class SoundboardSoundsUpdateEvent extends DispatchEvent {
 
   /// {@macro soundboard_sounds_update_event}
   /// @nodoc
-  SoundboardSoundsUpdateEvent({required super.gateway, required this.guildId, required this.sounds, required this.oldSounds});
+  SoundboardSoundsUpdateEvent({required super.client, required this.guildId, required this.sounds, required this.oldSounds});
 }
