+++ lib/src/utils/cache_helpers.dart
@@ -9,22 +9,22 @@ import 'package:nyxx/src/models/commands/application_command.dart';
 import 'package:nyxx/src/models/commands/application_command_permissions.dart';
 import 'package:nyxx/src/models/emoji.dart';
 import 'package:nyxx/src/models/entitlement.dart';
-import 'package:nyxx/src/models/gateway/event.dart';
-import 'package:nyxx/src/models/gateway/events/application_command.dart';
-import 'package:nyxx/src/models/gateway/events/auto_moderation.dart';
-import 'package:nyxx/src/models/gateway/events/channel.dart';
-import 'package:nyxx/src/models/gateway/events/entitlement.dart';
-import 'package:nyxx/src/models/gateway/events/guild.dart';
-import 'package:nyxx/src/models/gateway/events/integration.dart';
-import 'package:nyxx/src/models/gateway/events/interaction.dart';
-import 'package:nyxx/src/models/gateway/events/invite.dart';
-import 'package:nyxx/src/models/gateway/events/message.dart';
-import 'package:nyxx/src/models/gateway/events/presence.dart';
-import 'package:nyxx/src/models/gateway/events/ready.dart';
-import 'package:nyxx/src/models/gateway/events/soundboard.dart';
-import 'package:nyxx/src/models/gateway/events/stage_instance.dart';
-import 'package:nyxx/src/models/gateway/events/voice.dart';
-import 'package:nyxx/src/models/gateway/events/webhook.dart';
+import 'package:nyxx/src/models/events/event.dart';
+import 'package:nyxx/src/models/events/application_command.dart';
+import 'package:nyxx/src/models/events/auto_moderation.dart';
+import 'package:nyxx/src/models/events/channel.dart';
+import 'package:nyxx/src/models/events/entitlement.dart';
+import 'package:nyxx/src/models/events/guild.dart';
+import 'package:nyxx/src/models/events/integration.dart';
+import 'package:nyxx/src/models/events/interaction.dart';
+import 'package:nyxx/src/models/events/invite.dart';
+import 'package:nyxx/src/models/events/message.dart';
+import 'package:nyxx/src/models/events/presence.dart';
+import 'package:nyxx/src/models/events/ready.dart';
+import 'package:nyxx/src/models/events/soundboard.dart';
+import 'package:nyxx/src/models/events/stage_instance.dart';
+import 'package:nyxx/src/models/events/voice.dart';
+import 'package:nyxx/src/models/events/webhook.dart';
 import 'package:nyxx/src/models/guild/audit_log.dart';
 import 'package:nyxx/src/models/guild/auto_moderation.dart';
 import 'package:nyxx/src/models/guild/ban.dart';
